+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.4
 
 - Add tag parameter to To header if missing
+- Use header values for response host and port
 
 ## 0.3.3